+++ lib/src/models/footnote_update.dart
@@ -1,7 +1,7 @@
 /*
  * --------------------------------------------------------------------------------
  * <copyright company="Aspose" file="footnote_update.dart">
- *   Copyright (c) 2023 Aspose.Words for Cloud
+ *   Copyright (c) 2024 Aspose.Words for Cloud
  * </copyright>
  * <summary>
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -40,7 +40,7 @@ class FootnoteUpdate extends FootnoteBase {
 
     super.deserialize(json);
     if (json.containsKey('Position')) {
-      position = ModelBase.createInstance< NewDocumentPosition >(json['Position'] as Map<String, dynamic>);
+      position = ModelBase.createInstance< Position >(json['Position'] as Map<String, dynamic>);
     } else {
       position = null;
     }